remove MoveOnlyFunction and extract BitMap
