+++ test/cpp/dynamic_embedding/CMakeLists.txt
@@ -11,5 +11,4 @@ function(add_tde_test NAME)
 endfunction()
 
 add_tde_test(bits_op_test bits_op_test.cpp)
-add_tde_test(move_only_function_test move_only_function_test.cpp)
 add_tde_test(naive_id_transformer_test naive_id_transformer_test.cpp)
